+++ src/commands/index.ts
@@ -1,4 +1,5 @@
 export { registerInitSquadCommand } from './initSquadCommand';
+export { registerUpgradeSquadCommand } from './upgradeSquadCommand';
 export { registerAddMemberCommand } from './addMemberCommand';
 export { registerRemoveMemberCommand } from './removeMemberCommand';
 export { registerAddSkillCommand } from './addSkillCommand';